# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/playlists/me": {
      "get": {
        "operationId": "getMyPlaylists",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "includePrivate",
            "required": false,
            "in": "query",
            "schema": {
              "default": false,
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "onlyUserOwned",
            "required": false,
            "in": "query",
            "schema": {
              "default": false,
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get current user's playlists",
        "tags": [
          "Api"
        ]
      }
    },
    "/playlists/{id}": {
      "get": {
        "operationId": "getPlaylistById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Playlist not found"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get playlist by ID",
        "tags": [
          "Api"
        ]
      }
    },
    "/game/start": {
      "post": {
        "operationId": "startGame",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Start a new game from a playlist or daily",
        "tags": [
          "Api"
        ]
      }
    },
    "/game/history": {
      "get": {
        "operationId": "getHistory",
        "parameters": [
          {
            "name": "isDaily",
            "required": false,
            "in": "query",
            "description": "Whether the game is daily",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The limit of the history",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "The offset of the history",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameHistoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get user's game session history (paginated)",
        "tags": [
          "Api"
        ]
      }
    },
    "/game/stats": {
      "get": {
        "operationId": "getStats",
        "parameters": [
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "description": "The game mode",
            "schema": {
              "enum": [
                "DAILY",
                "ALL"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStatsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get user's daily stats",
        "tags": [
          "Api"
        ]
      }
    },
    "/game/daily/played-today": {
      "get": {
        "operationId": "getPlayedToday",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayedTodayDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Whether the user has played today's daily",
        "tags": [
          "Api"
        ]
      }
    },
    "/game/share/{id}": {
      "get": {
        "operationId": "getShare",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get shareable result for a game session",
        "tags": [
          "Api"
        ]
      }
    },
    "/game/{id}": {
      "get": {
        "operationId": "getGameState",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The internal Game Session UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get current game state",
        "tags": [
          "Api"
        ]
      }
    },
    "/game/{id}/guess": {
      "post": {
        "operationId": "submitGuess",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The internal Game Session UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuessResultDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuessResultDto"
                }
              }
            }
          }
        },
        "summary": "Submit a guess for a specific session",
        "tags": [
          "Api"
        ]
      }
    },
    "/search/tracks": {
      "get": {
        "operationId": "searchTracks",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Track options",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackOptionDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Search Spotify tracks (for game guess options)",
        "tags": [
          "Api"
        ]
      }
    },
...
  },
  "info": {
    "title": "Unpaused Spotify API",
    "description": "The backend for the Spotify guessing game",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "unpaused_session"
      }
    },
    "schemas": {
...
      "PlaylistDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the playlist"
          },
          "name": {
            "type": "string",
            "description": "The name of the playlist"
          },
          "description": {
            "type": "string",
            "description": "The description of the playlist",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "description": "The image URL of the playlist"
          },
          "owner": {
            "type": "string",
            "description": "The owner of the playlist"
          },
          "totalTracks": {
            "type": "number",
            "description": "The total number of tracks in the playlist"
          },
          "isPublic": {
            "type": "boolean",
            "description": "Whether the playlist is public"
          },
          "externalUrl": {
            "type": "string",
            "description": "The external URL of the playlist"
          }
        },
        "required": [
          "id",
          "name",
          "imageUrl",
          "owner",
          "totalTracks",
          "isPublic",
          "externalUrl"
        ]
      },
      "PlaylistsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "The playlists",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistDto"
            }
          },
          "total": {
            "type": "number",
            "description": "The total number of playlists"
          },
          "limit": {
            "type": "number",
            "description": "The limit of the playlists"
          },
          "offset": {
            "type": "number",
            "description": "The offset of the playlists"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "StartGameDto": {
        "type": "object",
        "properties": {
          "playlistId": {
            "type": "string",
            "description": "The ID of the playlist to start the game with"
          },
          "isDaily": {
            "type": "boolean",
            "description": "Whether the game is daily"
          }
        }
      },
      "GuessHistoryDto": {
        "type": "object",
        "properties": {
          "trackId": {
            "type": "string",
            "description": "The ID of the track"
          },
          "trackName": {
            "type": "string",
            "description": "The name of the track"
          },
          "artistName": {
            "type": "string",
            "description": "The artist of the track"
          },
          "result": {
            "enum": [
              "CORRECT",
              "ARTIST",
              "ALBUM",
              "WRONG",
              "SKIP",
              "ARTIST_AND_ALBUM"
            ],
            "type": "string",
            "description": "The result of the guess"
          }
        },
        "required": [
          "result"
        ]
      },
      "TrackOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the track"
          },
          "name": {
            "type": "string",
            "description": "The name of the track"
          },
          "normalizedName": {
            "type": "string",
            "description": "Normalized name for search/matching"
          },
          "artist": {
            "type": "string",
            "description": "The artist of the track"
          },
          "normalizedArtist": {
            "type": "string",
            "description": "Normalized artist for search/matching"
          },
          "albumImageUrl": {
            "type": "string",
            "description": "The album image URL of the track"
          },
          "albumName": {
            "type": "string",
            "description": "The album name of the track"
          },
          "albumUrl": {
            "type": "string",
            "description": "The album URL of the track"
          },
          "releaseYear": {
            "type": "number",
            "description": "The release year of the track"
          }
        },
        "required": [
          "id",
          "name",
          "normalizedName",
          "artist",
          "normalizedArtist"
        ]
      },
      "MetaGameExtrasVo": {
        "type": "object",
        "properties": {
          "winEmoji": {
            "type": "string",
            "description": "Show this emoji on the win screen"
          }
        },
        "required": [
          "winEmoji"
        ]
      },
      "GameStateDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "The ID of the game session"
          },
          "currentRound": {
            "type": "number",
            "description": "The current round of the game"
          },
          "snippetDuration": {
            "type": "number",
            "description": "The duration of the current snippet"
          },
          "status": {
            "type": "string",
            "description": "The status of the game",
            "enum": [
              "PLAYING",
              "WON",
              "LOST"
            ]
          },
          "guesses": {
            "description": "The guesses of the game",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuessHistoryDto"
            }
          },
          "previewUrl": {
            "type": "string",
            "description": "The preview URL of the current track"
          },
          "answer": {
            "description": "The answer to the current track",
            "allOf": [
              {
                "$ref": "#/components/schemas/TrackOptionDto"
              }
            ]
          },
          "rankTitle": {
            "type": "string",
            "description": "Optional personalized rank title (easter egg for special users)"
          },
          "specialNote": {
            "type": "string",
            "description": "Optional personalized note (easter egg for special users)"
          },
          "meta": {
            "description": "Optional win easter eggs",
            "allOf": [
              {
                "$ref": "#/components/schemas/MetaGameExtrasVo"
              }
            ]
          }
        },
        "required": [
          "sessionId",
          "currentRound",
          "snippetDuration",
          "status",
          "guesses",
          "previewUrl"
        ]
      },
      "GameHistoryEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Game session ID"
          },
          "date": {
            "type": "string",
            "description": "Date (YYYY-MM-DD) from completedAt or createdAt"
          },
          "status": {
            "type": "string",
            "description": "Game status",
            "enum": [
              "PLAYING",
              "WON",
              "LOST"
            ]
          },
          "score": {
            "type": "number",
            "description": "Score 0-6 if completed"
          },
          "isDaily": {
            "type": "boolean",
            "description": "Whether the game was daily"
          },
          "guesses": {
            "description": "Guess history",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuessHistoryDto"
            }
          },
          "trackName": {
            "type": "string",
            "description": "Track name (answer)"
          },
          "artistName": {
            "type": "string",
            "description": "Artist name"
          },
          "albumImageUrl": {
            "type": "string",
            "description": "Album image URL"
          },
          "playlistName": {
            "type": "string",
            "description": "Playlist/source name e.g. Liked Songs"
          }
        },
        "required": [
          "id",
          "date",
          "status",
          "isDaily",
          "guesses",
          "trackName",
          "artistName"
        ]
      },
      "GameHistoryDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameHistoryEntryDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total count"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "GameStatsDto": {
        "type": "object",
        "properties": {
          "currentStreak": {
            "type": "number",
            "description": "The current streak"
          },
          "bestStreak": {
            "type": "number",
            "description": "The best streak"
          },
          "totalGames": {
            "type": "number",
            "description": "The total games"
          },
          "totalWins": {
            "type": "number",
            "description": "The total wins"
          },
          "roundDistribution": {
            "description": "The round distribution",
            "example": [
              0,
              0,
              0,
              0,
              0,
              0,
              0
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "mode": {
            "type": "string",
            "description": "The game mode",
            "enum": [
              "DAILY",
              "ALL"
            ]
          },
          "winRate": {
            "type": "number",
            "description": "The win rate",
            "example": 0.5
          },
          "averageScore": {
            "type": "number",
            "description": "The average score",
            "example": 3.5
          }
        },
        "required": [
          "currentStreak",
          "bestStreak",
          "totalGames",
          "totalWins",
          "roundDistribution",
          "mode",
          "winRate",
          "averageScore"
        ]
      },
      "PlayedTodayDto": {
        "type": "object",
        "properties": {
          "playedToday": {
            "type": "boolean",
            "description": "Whether the user has played today's daily"
          }
        },
        "required": [
          "playedToday"
        ]
      },
      "ShareResultDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Date (YYYY-MM-DD)"
          },
          "score": {
            "type": "number",
            "description": "Score 0-6"
          },
          "guessPattern": {
            "type": "string",
            "description": "Guess pattern e.g. 🟥🟨🟩"
          },
          "trackName": {
            "type": "string",
            "description": "Track name"
          },
          "artistName": {
            "type": "string",
            "description": "Artist name"
          },
          "shareText": {
            "type": "string",
            "description": "Pre-formatted share text"
          },
          "gameNumber": {
            "type": "number",
            "description": "Game number (days since epoch or similar)"
          }
        },
        "required": [
          "date",
          "score",
          "guessPattern",
          "trackName",
          "artistName",
          "shareText",
          "gameNumber"
        ]
      },
      "GuessDto": {
        "type": "object",
        "properties": {
          "trackId": {
            "type": "string",
            "description": "The ID of the track to guess"
          },
          "skip": {
            "type": "boolean",
            "default": false,
            "description": "Whether the current guess is a skip"
          },
          "trackName": {
            "type": "string",
            "description": "Track name"
          },
          "artistName": {
            "type": "string",
            "description": "Artist name"
          },
          "albumName": {
            "type": "string",
            "description": "Album name"
          }
        }
      },
      "GuessResultDto": {
        "type": "object",
        "properties": {
          "result": {
            "enum": [
              "CORRECT",
              "ARTIST",
              "ALBUM",
              "WRONG",
              "SKIP",
              "ARTIST_AND_ALBUM"
            ],
            "type": "string",
            "description": "The result of the guess"
          },
          "gameOver": {
            "type": "boolean",
            "description": "Whether the game is over"
          },
          "status": {
            "type": "string",
            "description": "The status of the game",
            "enum": [
              "PLAYING",
              "WON",
              "LOST"
            ]
          },
          "currentRound": {
            "type": "number",
            "description": "The current round of the game"
          },
          "snippetDuration": {
            "type": "number",
            "description": "The duration of the current snippet"
          },
          "rankTitle": {
            "type": "string",
            "description": "Optional personalized rank title (easter egg)"
          },
          "specialNote": {
            "type": "string",
            "description": "Optional personalized note (easter egg)"
          },
          "meta": {
            "type": "object",
            "description": "Optional win easter eggs"
          }
        },
        "required": [
          "result",
          "gameOver",
          "status",
          "currentRound",
          "snippetDuration"
        ]
      },
...
    }
  }
}